+++ src/flush.c
@@ -20,8 +20,7 @@
 #include "log_internal.h"
 #include "mr.h"
 
-static int rpma_flush_apm_new(struct rpma_peer *peer,
-		struct rpma_flush *flush);
+static int rpma_flush_apm_new(struct rpma_peer *peer, struct rpma_flush *flush);
 static int rpma_flush_apm_delete(struct rpma_flush *flush);
 static int rpma_flush_apm_execute(struct ibv_qp *qp, struct rpma_flush *flush,
 	struct rpma_mr_remote *dst, size_t dst_offset, size_t len,
@@ -62,16 +61,15 @@ rpma_flush_apm_new(struct rpma_peer *peer, struct rpma_flush *flush)
 	/* a memory registration has to be page-aligned */
 	long pagesize = sysconf(_SC_PAGESIZE);
 	if (pagesize < 0) {
-		RPMA_LOG_FATAL("sysconf(_SC_PAGESIZE) failed: %s",
-				strerror(errno));
+		RPMA_LOG_FATAL("sysconf(_SC_PAGESIZE) failed: %s", strerror(errno));
 		return RPMA_E_PROVIDER;
 	}
 
 	size_t mmap_size = (size_t)pagesize;
 
 	/* allocate memory for the read-after-write buffer (RAW) */
-	void *raw = mmap(NULL, mmap_size, PROT_READ | PROT_WRITE,
-			MAP_SHARED | MAP_ANONYMOUS, -1, 0);
+	void *raw = mmap(NULL, mmap_size, PROT_READ | PROT_WRITE, MAP_SHARED | MAP_ANONYMOUS,
+			-1, 0);
 	if (raw == MAP_FAILED)
 		return RPMA_E_NOMEM;
 
@@ -94,8 +92,7 @@ rpma_flush_apm_new(struct rpma_peer *peer, struct rpma_flush *flush)
 	flush_apm->raw_mr = raw_mr;
 	flush_apm->mmap_size = mmap_size;
 
-	struct rpma_flush_internal *flush_internal =
-			(struct rpma_flush_internal *)flush;
+	struct rpma_flush_internal *flush_internal = (struct rpma_flush_internal *)flush;
 	flush_internal->flush_func = rpma_flush_apm_execute;
 	flush_internal->delete_func = rpma_flush_apm_delete;
 	flush_internal->context = flush_apm;
@@ -111,10 +108,8 @@ rpma_flush_apm_delete(struct rpma_flush *flush)
 {
 	RPMA_DEBUG_TRACE;
 
-	struct rpma_flush_internal *flush_internal =
-			(struct rpma_flush_internal *)flush;
-	struct flush_apm *flush_apm =
-			(struct flush_apm *)flush_internal->context;
+	struct rpma_flush_internal *flush_internal = (struct rpma_flush_internal *)flush;
+	struct flush_apm *flush_apm = (struct flush_apm *)flush_internal->context;
 
 	int ret_dereg = rpma_mr_dereg(&flush_apm->raw_mr);
 	int ret_unmap = munmap(flush_apm->raw, flush_apm->mmap_size);
@@ -134,20 +129,18 @@ rpma_flush_apm_delete(struct rpma_flush *flush)
  * rpma_flush_apm_execute -- perform the APM-style flush
  */
 static int
-rpma_flush_apm_execute(struct ibv_qp *qp, struct rpma_flush *flush,
-	struct rpma_mr_remote *dst, size_t dst_offset, size_t len,
-	enum rpma_flush_type type, int flags, const void *op_context)
+rpma_flush_apm_execute(struct ibv_qp *qp, struct rpma_flush *flush, struct rpma_mr_remote *dst,
+	size_t dst_offset, size_t len, enum rpma_flush_type type, int flags,
+	const void *op_context)
 {
 	RPMA_DEBUG_TRACE;
 	RPMA_FAULT_INJECTION(RPMA_E_PROVIDER, {});
 
-	struct rpma_flush_internal *flush_internal =
-			(struct rpma_flush_internal *)flush;
-	struct flush_apm *flush_apm =
-			(struct flush_apm *)flush_internal->context;
+	struct rpma_flush_internal *flush_internal = (struct rpma_flush_internal *)flush;
+	struct flush_apm *flush_apm = (struct flush_apm *)flush_internal->context;
 
-	return rpma_mr_read(qp, flush_apm->raw_mr, 0, dst, dst_offset,
-			RAW_SIZE, flags, op_context);
+	return rpma_mr_read(qp, flush_apm->raw_mr, 0, dst, dst_offset, RAW_SIZE, flags,
+			op_context);
 }
 
 /* internal librpma API */
@@ -184,8 +177,7 @@ rpma_flush_delete(struct rpma_flush **flush_ptr)
 {
 	RPMA_DEBUG_TRACE;
 
-	struct rpma_flush_internal *flush_internal =
-			*(struct rpma_flush_internal **)flush_ptr;
+	struct rpma_flush_internal *flush_internal = *(struct rpma_flush_internal **)flush_ptr;
 
 	int ret = flush_internal->delete_func(*flush_ptr);
 	free(*flush_ptr);