+++ src/peer.h
@@ -34,9 +34,8 @@ int rpma_peer_create_srq(struct rpma_peer *peer, struct rpma_srq_cfg *cfg,
  * - RPMA_E_INVAL - peer, id or cq is NULL
  * - RPMA_E_PROVIDER - allocating a QP failed
  */
-int rpma_peer_setup_qp(struct rpma_peer *peer, struct rdma_cm_id *id,
-		struct rpma_cq *cq, struct rpma_cq *rcq,
-		const struct rpma_conn_cfg *cfg);
+int rpma_peer_setup_qp(struct rpma_peer *peer, struct rdma_cm_id *id, struct rpma_cq *cq,
+		struct rpma_cq *rcq, const struct rpma_conn_cfg *cfg);
 
 /*
  * ASSUMPTIONS
@@ -48,7 +47,7 @@ int rpma_peer_setup_qp(struct rpma_peer *peer, struct rdma_cm_id *id,
  *
  * - RPMA_E_PROVIDER - registering the memory region failed
  */
-int rpma_peer_setup_mr_reg(struct rpma_peer *peer, struct ibv_mr **ibv_mr_ptr,
-		void *addr, size_t length, int usage);
+int rpma_peer_setup_mr_reg(struct rpma_peer *peer, struct ibv_mr **ibv_mr_ptr, void *addr,
+		size_t length, int usage);
 
 #endif /* LIBRPMA_PEER_H */