+++ CHANGELOG.md
@@ -67,6 +67,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - CMAKE_BUILD_TYPE from Debug to Release
   - TESTS_PERF_TOOLS - from ON to OFF
 
+ - the new line length is 100 characters
+
 ### Fixed
 - APIs:
   - rpma_peer_delete - fixed memory leak when ibv_dealloc_pd() fails